Mavenize: test work
- move additional test files around

Fixes: https://github.com/unicode-org/unicodetools/issues/64
X-PR-URL: https://github.com/unicode-org/unicodetools/pull/91
