+++ unicodetools/src/main/java/org/unicode/text/tools/Collections.java
@@ -1,0 +1,5 @@
+package org.unicode.text.tools;
+
+public class Collections {
+
+}
