+++ unicodetools/src/test/java/org/unicode/draft/Test3.java
