+++ unicodetools/src/test/java/org/unicode/draft/TestCompressed.java
