+++ unicodetools/src/test/java/org/unicode/propstest/CheckEmojiProps.java
