+++ unicodetools/src/test/java/org/unicode/tools/emoji/unittest/Asserts.java
@@ -1,15 +1,15 @@
 package org.unicode.tools.emoji.unittest;
 
+import com.ibm.icu.text.UnicodeSet;
+
 import org.unicode.cldr.util.UnicodeSetPrettyPrinter;
 import org.unicode.text.utility.Utility;
-
-import com.ibm.icu.dev.test.TestLog;
-import com.ibm.icu.text.UnicodeSet;
+import org.unicode.unittest.TestFmwkMinusMinus;
 
 public class Asserts {
     // TODO move these into common place
-    
-    public static void assertEqualsUS(TestLog log, String message, String s1Name, UnicodeSet s1, String s2Name, UnicodeSet s2) {
+
+    public static void assertEqualsUS(TestFmwkMinusMinus log, String message, String s1Name, UnicodeSet s1, String s2Name, UnicodeSet s2) {
         if (s1.equals(s2)) {
             return;
         }
@@ -17,14 +17,14 @@ public class Asserts {
         assertContains(log, message, s2Name, s2, s1Name, s1);
     }
 
-    public static void assertContains(TestLog log, String message, String s1Name, UnicodeSet s1, String s2Name, UnicodeSet s2) {
+    public static void assertContains(TestFmwkMinusMinus log, String message, String s1Name, UnicodeSet s1, String s2Name, UnicodeSet s2) {
         if (!s1.containsAll(s2)) {
             UnicodeSet s2minuss1 = new UnicodeSet(s2).removeAll(s1);
             log.errln(message + ", " + s1Name + " ⊉ " + s2Name + ": " + s2minuss1.toPattern(false) + " ≠ ∅");
         }
     }
-    
-    public static void assertContains(TestLog log, String message, String s1Name, UnicodeSet s1, String s2Name, String s2char) {
+
+    public static void assertContains(TestFmwkMinusMinus log, String message, String s1Name, UnicodeSet s1, String s2Name, String s2char) {
         if (!s1.contains(s2char)) {
             log.errln(message + ", " + s1Name + " ∌ " + s2Name + ": " + s2char + ", " + Utility.hex(s2char));
         }
