+++ unicodetools/src/test/java/org/unicode/utilities/TestPolaritySet.java
